# perf: improve url components getter